feat: pdf now loads on home screen
