+++ lib/ui/views/home/home_loading.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-class PdfError extends StatelessWidget {
+class HomeLoading extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Center(
-      child: Text('Error loading PDF\nWOMP WOMP!'),
+      child: CircularProgressIndicator(),
     );
   }
 }
