--- lib/ui/views/home/pdf_loading.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PdfLoading extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: CircularProgressIndicator(),
-    );
-  }
-}
