+++ lib/ui/views/views.dart
@@ -1,2 +1,1 @@
 export 'home/home_screen.dart';
-export 'pdf/pdf_screen.dart';
